Add support for multi-projects
In order to run gradle with a multi-project configuration, we need to
extract all the build.gradle and settings.gradle configurations before
running any gradle command. In order to do that, I've created a
preExtract method that will be called before the extractDependencies
so we can use it to save all the required files.